pyproject.toml
[tool.poetry]
name = "self-organizing-trading-agents"
version = "0.1.0"
description = "Self-organizing multi-agent AI trading system with options strategies"
authors = ["AI Trading Team"]
readme = "README.md"
packages = [{include = "agents"}, {include = "strategies"}, {include = "risk"}, {include = "exec"}, {include = "data_providers"}, {include = "storage"}, {include = "graphs"}]

[tool.poetry.dependencies]
python = "^3.10"
langchain = "^0.1.0"
langgraph = "^0.0.20"
crewai = "^0.1.0"
numpy = "^1.24.0"
pandas = "^2.0.0"
yfinance = "^0.2.28"
requests = "^2.31.0"
python-dotenv = "^1.0.0"
pyyaml = "^6.0"
sqlalchemy = "^2.0.0"
chromadb = "^0.4.0"
streamlit = "^1.28.0"
plotly = "^5.17.0"
scipy = "^1.11.0"
scikit-learn = "^1.3.0"
pydantic = "^2.0.0"
openai = "^1.3.0"
anthropic = "^0.8.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
black = "^23.7.0"
ruff = "^0.0.285"
mypy = "^1.5.0"
jupyter = "^1.0.0"
ipykernel = "^6.25.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --cov=. --cov-report=html --cov-report=term"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
